Moved instantiation into a single file
