+++ nevergrad/instrumentation/__init__.py
@@ -3,8 +3,7 @@
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
 
-from .folderfunction import FolderFunction
-from .instantiate import InstrumentedFunction, register_file_type
+from .instantiate import register_file_type, FolderFunction
 from . import variables
-from .variables import Instrumentation
+from .variables import Instrumentation, InstrumentedFunction
 from .utils import TemporaryDirectoryCopy
