+++ nevergrad/instrumentation/test_instantiate.py
@@ -6,14 +6,13 @@
 import tempfile
 from pathlib import Path
 from unittest import TestCase
-from typing import Tuple, Dict, Any
+from typing import Tuple, List, Optional
 import genty
 import numpy as np
 from ..common import testing
 from . import instantiate
 from .instantiate import LINETOKEN
-from . import variables
-from . import utils
+from .instantiate import Placeholder
 
 
 def test_symlink_folder_tree() -> None:
@@ -60,76 +59,46 @@ class InstantiationTests(TestCase):
         without_clean_copy=(False,),
     )
     def test_folder_instantiator(self, clean_copy: bool) -> None:
-        path = Path(__file__).parent / "examples" / "basic"
+        path = Path(__file__).parent / "examples"
         ifolder = instantiate.FolderInstantiator(path, clean_copy=clean_copy)
-        testing.printed_assert_equal(ifolder.placeholders, [utils.Placeholder("value1", "this is a comment"),
-                                                            utils.Placeholder("value2", None)])
+        testing.printed_assert_equal(ifolder.placeholders, [Placeholder("value1", "this is a comment"),
+                                                            Placeholder("value2", None)])
         np.testing.assert_equal(len(ifolder.file_functions), 1)
         with ifolder.instantiate(value1=12, value2=110.) as tmp:
             with (tmp / "script.py").open("r") as f:
                 lines = f.readlines()
         np.testing.assert_equal(lines[10], "value2 = 110.0\n")
 
+    @genty.genty_dataset(  # type: ignore
+        void=("bvcebsl\nsoefn", []),
+        unique_no_comment=("bfseibf\nbsfei NG_VAR{machin}", [("machin", None)]),
+        several=("bfkes\nsgrdgrgbdrkNG_VAR{truc|blublu}sehnNG_VAR{bidule}", [("truc", "blublu"), ("bidule", None)]),
+    )
+    def test_placeholder(self, text: str, name_comments: List[Tuple[str, Optional[str]]]) -> None:
+        placeholders = Placeholder.finditer(text)
+        testing.printed_assert_equal(placeholders, [Placeholder(*x) for x in name_comments])
+
+
+def test_placeholder_substitution() -> None:
+    text = "bfkes\nsgrdgrgbdrkNG_VAR{truc|blublu}sehn NG_VAR{bidule}"
+    expected = "bfkes\nsgrdgrgbdrk#12#sehn 24"
+    output = Placeholder.sub(text, truc="#12#", bidule=24)
+    np.testing.assert_equal(output, expected)
+    np.testing.assert_raises(KeyError, Placeholder.sub, text, truc="#12#")
+    np.testing.assert_raises(RuntimeError, Placeholder.sub, text, truc="#12#", bidule=24, chouette=12)
+    text = "bfkes\nsgrdgrgbdrkNG_VAR{truc|blublu}sehnNG_VAR{bidule}NG_VAR{bidule|bis}"
+    np.testing.assert_raises(RuntimeError, Placeholder.sub, text, truc="#12#", bidule=24)
+
 
 def test_file_text_function() -> None:
-    path = Path(__file__).parent / "examples" / "basic" / "script.py"
+    path = Path(__file__).parent / "examples" / "script.py"
     filefunc = instantiate.FileTextFunction(path)
-    testing.printed_assert_equal(filefunc.placeholders, [utils.Placeholder("value1", "this is a comment"),
-                                                         utils.Placeholder("value2", None)])
+    testing.printed_assert_equal(filefunc.placeholders, [Placeholder("value1", "this is a comment"),
+                                                         Placeholder("value2", None)])
 
 
-def _arg_return(*args: Any, **kwargs: Any) -> Tuple[Tuple[Any, ...], Dict[str, Any]]:
-    return args, kwargs
-
-
-def test_instrumented_function() -> None:
-    ifunc = instantiate.InstrumentedFunction(_arg_return, variables.SoftmaxCategorical([1, 12]), "constant",
-                                             variables.Gaussian(0, 1, [2, 2]), constkwarg="blublu",
-                                             plop=variables.SoftmaxCategorical([3, 4]))
-    np.testing.assert_equal(ifunc.dimension, 8)
-    data = [-100, 100, 1, 2, 3, 4, 100, -100]
-    args, kwargs = ifunc(data)
-    testing.printed_assert_equal(args, [12, "constant", [[1, 2], [3, 4]]])
-    testing.printed_assert_equal(kwargs, {"constkwarg": "blublu", "plop": 3})
-    testing.printed_assert_equal(ifunc.descriptors, {"dimension": 8, "name": "_arg_return", "instrumented": "arg0,arg2,plop",
-                                                     "function_class": "InstrumentedFunction", "transform": None})
-    print(ifunc.get_summary(data))
-
-
-def test_instrumented_function_kwarg_order() -> None:
-    ifunc = instantiate.InstrumentedFunction(_arg_return, kw4=variables.SoftmaxCategorical([1, 0]), kw2="constant",
-                                             kw3=variables.Gaussian(0, 1, [2, 2]), kw1=variables.Gaussian(2, 2))
-    np.testing.assert_equal(ifunc.dimension, 7)
-    data = [-1, 1, 2, 3, 4, 100, -100]
-    _, kwargs = ifunc(data)
-    testing.printed_assert_equal(kwargs, {"kw1": 0, "kw2": "constant", "kw3": [[1, 2], [3, 4]], "kw4": 1})
-
-
-class _Callable:
-
-    def __call__(self, x: float, y: float = 0) -> float:
-        return abs(x + y)
-
-
-def test_callable_instrumentation() -> None:
-    ifunc = instantiate.InstrumentedFunction(lambda x: x**2, variables.Gaussian(2, 2))
-    np.testing.assert_equal(ifunc.descriptors["name"], "<lambda>")
-    ifunc = instantiate.InstrumentedFunction(_Callable(), variables.Gaussian(2, 2))
-    np.testing.assert_equal(ifunc.descriptors["name"], "_Callable")
-
-
-def test_deterministic_convert_to_args() -> None:
-    ifunc = instantiate.InstrumentedFunction(_Callable(), variables.SoftmaxCategorical([0, 1, 2, 3]),
-                                             y=variables.SoftmaxCategorical([0, 1, 2, 3]))
-    data = [.01, 0, 0, 0, .01, 0, 0, 0]
-    for _ in range(20):
-        args, kwargs = ifunc.convert_to_arguments(data, deterministic=True)
-        testing.printed_assert_equal(args, [0])
-        testing.printed_assert_equal(kwargs, {"y": 0})
-    arg_sum, kwarg_sum = 0, 0
-    for _ in range(24):
-        args, kwargs = ifunc.convert_to_arguments(data, deterministic=False)
-        arg_sum += args[0]
-        kwarg_sum += kwargs["y"]
-    assert arg_sum != 0
-    assert kwarg_sum != 0
+def test_folder_function() -> None:
+    folder = Path(__file__).parent / "examples"
+    func = instantiate.FolderFunction(str(folder), ["python", "examples/script.py"], clean_copy=True)
+    output = func(value1=98, value2=6)
+    np.testing.assert_equal(output, 12)
