+++ nevergrad/instrumentation/variables.py
@@ -4,9 +4,10 @@
 # LICENSE file in the root directory of this source tree.
 
 import itertools
-from typing import List, Any, Match, Optional, Tuple, Dict
+from typing import List, Any, Match, Optional, Tuple, Dict, Callable
 import numpy as np
 from . import discretization
+from ..functions import base
 from ..common.typetools import ArrayLike
 from . import utils
 
@@ -215,3 +216,79 @@ class Instrumentation:
                                      f"but expected:\n{self.names}")
         data = list(itertools.chain.from_iterable([instrument.process_arg(arg) for instrument, arg in zip(self.instruments, arguments)]))
         return np.array(data)
+
+
+class InstrumentedFunction(base.BaseFunction):
+    """Converts a multi-argument function into a mono-argument multidimensional continuous function
+    which can be optimized.
+
+    Parameters
+    ----------
+    function: callable
+        the callable to convert
+    *args, **kwargs: Any
+        Any argument. Arguments of type variables.SoftmaxCategorical or variables.Gaussian will be instrumented
+        and others will be kept constant.
+
+    Note
+    ----
+    - Tokens can be:
+      - DiscreteToken(list_of_n_possible_values): converted into a n-dim array, corresponding to proba for each value
+      - GaussianToken(mean, std, shape=None): a Gaussian variable (shape=None) or array.
+    - This function can then be directly used in benchmarks *if it returns a float*.
+
+    """
+
+    def __init__(self, function: Callable, *args: Any, **kwargs: Any) -> None:
+        assert callable(function)
+        self.instrumentation = Instrumentation(*args, **kwargs)
+        super().__init__(dimension=self.instrumentation.dimension)
+        # keep track of what is instrumented (but "how" is probably too long/complex)
+        instrumented = [f"arg{k}" if name is None else name for k, name in enumerate(self.instrumentation.names)
+                        if not isinstance(self.instrumentation.instruments[k], _Constant)]
+        name = function.__name__ if hasattr(function, "__name__") else function.__class__.__name__
+        self._descriptors.update(name=name, instrumented=",".join(instrumented))
+        self._function = function
+        self.last_call_args: Optional[Tuple[Any, ...]] = None
+        self.last_call_kwargs: Optional[Dict[str, Any]] = None
+
+    def convert_to_arguments(self, data: np.ndarray, deterministic: bool = True) -> Tuple[Tuple[Any, ...], Dict[str, Any]]:
+        """Get the arguments and keyword arguments corresponding to the data
+
+        Parameters
+        ----------
+        data: np.ndarray
+            input data
+        deterministic: bool
+            whether to process the data deterministically (some Variables such as SoftmaxCategorical are stochastic).
+            If True, the output is the most likely output.
+        """
+        return self.instrumentation.data_to_arguments(data, deterministic=deterministic)
+
+    def convert_to_data(self, *args: Any, **kwargs: Any) -> ArrayLike:
+        return self.instrumentation.arguments_to_data(*args, **kwargs)
+
+    def oracle_call(self, x: np.ndarray) -> Any:
+        self.last_call_args, self.last_call_kwargs = self.convert_to_arguments(x, deterministic=False)
+        return self._function(*self.last_call_args, **self.last_call_kwargs)
+
+    def __call__(self, x: np.ndarray) -> Any:
+        # BaseFunction __call__ method should generally not be overriden,
+        # but here that would mess up with typing, and I would rather not constrain
+        # user to return only floats.
+        x = self.transform(x)
+        return self.oracle_call(x)
+
+    def get_summary(self, data: np.ndarray) -> Any:  # probably impractical for large arrays
+        """Prints the summary corresponding to the provided data
+        """
+        strings = []
+        names = self.instrumentation.names
+        instruments = self.instrumentation.instruments
+        splitted_data = utils.split_data(data, instruments)
+        for k, (name, var, d) in enumerate(zip(names, instruments, splitted_data)):
+            if not isinstance(var, _Constant):
+                explanation = var.get_summary(d)
+                sname = f"arg #{k + 1}" if name is None else f'kwarg "{name}"'
+                strings.append(f"{sname}: {explanation}")
+        return " - " + "\n - ".join(strings)
